Update spec as if PR #645 was accepted
This change updates the formatting spec to reflect the
changes proposed in #645. It should not be merged as-is.
It also uses slightly different terms than the design doc in #645,
but should serve to give a sense of what the spec would look
like if the "composability" design doc was accepted.